[hall] removed averaging + shrunk variables that could be shrunk

+++ magnetic_sensor.h
@@ -7,7 +7,7 @@
 
 #include "spi_utils.h"
 
-uint32_t get_round_dt();
+uint16_t get_round_dt();
 uint32_t get_round_time();
 void magnetic_sensor_init();
